Update to use actions
