+++ app/(auth)/logout/LogoutButton.tsx
@@ -1,0 +1,25 @@
+'use client';
+import { useRouter } from 'next/navigation';
+import { useTransition } from 'react';
+import setCookie from './logoutAction';
+import styles from './LogoutButton.module.scss';
+
+export default function LogoutButtonClient() {
+  const [, startTransition] = useTransition();
+  const router = useRouter();
+
+  return (
+    <button
+      className={styles.logoutButton}
+      onClick={() => {
+        startTransition(() => {
+          setCookie().catch(() => {});
+        });
+        router.replace('/');
+        router.refresh();
+      }}
+    >
+      logout
+    </button>
+  );
+}
