--- app/(auth)/logout/revalidate/page.tsx
@@ -1,14 +1,0 @@
-'use client';
-
-import { useRouter } from 'next/navigation';
-import { useEffect } from 'react';
-
-export default function LogoutButtonClient() {
-  const router = useRouter();
-
-  useEffect(() => {
-    router.replace(`/`);
-    router.refresh();
-  });
-  return null;
-}
