+++ package.json
@@ -17,7 +17,7 @@
     "dotenv-safe": "^8.2.0",
     "js-cookie": "^3.0.1",
     "ley": "^0.8.1",
-    "next": "13.4.2-canary.3",
+    "next": "13.4.2",
     "postgres": "^3.3.3",
     "react": "18.2.0",
     "react-dom": "18.2.0",
